requirements.txt
numpy==2.2.3
pygame==2.6.1
pillow==11.1.0
emoji==2.14.1